+++ app/controllers/api/v1/sheets_controller.rb
@@ -1,34 +1,41 @@
 class Api::V1::SheetsController < ApplicationController
     def index
-
-      render json: Sheet.all
+      render json: current_user.sheets
     end
 
     def show
-
-      render json: Sheet.find(params[:id])
+      sheet = Sheet.find(params[:id])
+      if current_user.sheets.include? sheet
+        render json: sheet
+      else
+        render nothing: true
+      end
     end
 
     def create
-
       sheet = Sheet.new(sheet_params)
       sheet.save
       render json: sheet
     end
 
     def update
-
       sheet = Sheet.find(params[:id])
-      sheet.update(sheet_params)
-      render json: sheet
+      if current_user.sheets.include? sheet
+        sheet.update(sheet_params)
+        render json: sheet
+      else
+        render nothing: true
+      end
     end
 
     def destroy
-
       sheet = Sheet.find(params[:id])
-      sheet.destroy
+      if current_user.sheets.include? sheet
+        sheet.destroy
+      end
       render nothing: true
     end
+
 
 
     private
